Revert "Support for context documents in search (Personalization) (#287)"
This reverts commit 44a8d4850d77447ace4f060ca3f70de57f9d983a.
